refactor: move fucntion
